# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"每日英语听力 · 内容后台","description":"每日英语听力 App 内容后台 API。每天抓取多主题新闻，经 AI 改写为 5 个英语等级（CEFR A1–C1）与 5 个日语等级（JLPT N5–N1），并生成带逐词时间轴的免费 TTS 音频。","version":"1.0.0"},"paths":{"/health":{"get":{"tags":["meta"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthOut"}}}}}}},"/api/v1/levels":{"get":{"tags":["meta"],"summary":"List Levels","operationId":"list_levels_api_v1_levels_get","parameters":[{"name":"lang","in":"query","required":false,"schema":{"anyOf":[{"enum":["en","ja"],"type":"string"},{"type":"null"}],"description":"只返回该语言","title":"Lang"},"description":"只返回该语言"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LevelOut"},"title":"Response List Levels Api V1 Levels Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/topics":{"get":{"tags":["meta"],"summary":"List Topics","operationId":"list_topics_api_v1_topics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"type":"string"},"type":"array","title":"Response List Topics Api V1 Topics Get"}}}}}}},"/api/v1/articles":{"get":{"tags":["content"],"summary":"List Articles","operationId":"list_articles_api_v1_articles_get","parameters":[{"name":"date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"按发布日过滤 YYYY-MM-DD","title":"Date"},"description":"按发布日过滤 YYYY-MM-DD"},{"name":"topic","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Topic"}},{"name":"source","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"}},{"name":"lang","in":"query","required":false,"schema":{"anyOf":[{"enum":["en","ja"],"type":"string"},{"type":"null"}],"title":"Lang"}},{"name":"level","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":5,"minimum":1},{"type":"null"}],"title":"Level"}},{"name":"level_code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Level Code"}},{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"标题关键词","title":"Q"},"description":"标题关键词"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/articles/today":{"get":{"tags":["content"],"summary":"Today","description":"App 首页最常用的一次调用：今天 + 指定语言等级。","operationId":"today_api_v1_articles_today_get","parameters":[{"name":"lang","in":"query","required":false,"schema":{"enum":["en","ja"],"type":"string","default":"en","title":"Lang"}},{"name":"level","in":"query","required":false,"schema":{"type":"integer","maximum":5,"minimum":1,"default":1,"title":"Level"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":30,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/versions/{version_id}":{"get":{"tags":["content"],"summary":"Get Version","operationId":"get_version_api_v1_versions__version_id__get","parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleDetailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/articles/{article_id}":{"get":{"tags":["content"],"summary":"Get Article","description":"按「文章 id + 语言 + 等级」取内容，App 端从列表跳转详情用这个。","operationId":"get_article_api_v1_articles__article_id__get","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","title":"Article Id"}},{"name":"lang","in":"query","required":false,"schema":{"enum":["en","ja"],"type":"string","default":"en","title":"Lang"}},{"name":"level","in":"query","required":false,"schema":{"type":"integer","maximum":5,"minimum":1,"default":1,"title":"Level"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleDetailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/articles/{article_id}/levels":{"get":{"tags":["content"],"summary":"Article Levels","operationId":"article_levels_api_v1_articles__article_id__levels_get","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","title":"Article Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LevelOut"},"title":"Response Article Levels Api V1 Articles  Article Id  Levels Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/versions/{version_id}/audio":{"get":{"tags":["audio"],"summary":"Version Audio","operationId":"version_audio_api_v1_versions__version_id__audio_get","parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AudioOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audio/{audio_id}.mp3":{"get":{"tags":["audio"],"summary":"Serve Audio","description":"返回音频文件，支持 Range（播放器拖动进度条必需）。","operationId":"serve_audio_api_v1_audio__audio_id__mp3_get","parameters":[{"name":"audio_id","in":"path","required":true,"schema":{"type":"integer","title":"Audio Id"}},{"name":"download","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Download"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/versions/{version_id}/timeline":{"get":{"tags":["audio"],"summary":"Version Timeline","description":"逐词时间轴 —— App 跟读高亮的核心数据。","operationId":"version_timeline_api_v1_versions__version_id__timeline_get","parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}},{"name":"format","in":"query","required":false,"schema":{"enum":["compact","full"],"type":"string","description":"compact=精简字段，full=含原始 100ns boundary","default":"compact","title":"Format"},"description":"compact=精简字段，full=含原始 100ns boundary"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/versions/{version_id}/subtitle.srt":{"get":{"tags":["audio"],"summary":"Version Srt","operationId":"version_srt_api_v1_versions__version_id__subtitle_srt_get","parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs":{"get":{"tags":["jobs"],"summary":"List Jobs","operationId":"list_jobs_api_v1_jobs_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JobOut"},"title":"Response List Jobs Api V1 Jobs Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs/daily":{"post":{"tags":["jobs"],"summary":"手动触发每日抓取+改写+TTS","operationId":"trigger_daily_api_v1_jobs_daily_post","parameters":[{"name":"topics","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"逗号分隔，默认取配置全部","title":"Topics"},"description":"逗号分隔，默认取配置全部"},{"name":"per_topic","in":"query","required":false,"schema":{"type":"integer","maximum":5,"minimum":1,"default":1,"title":"Per Topic"}},{"name":"with_tts","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"With Tts"}},{"name":"force","in":"query","required":false,"schema":{"type":"boolean","description":"True=忽略已有结果重新生成","default":false,"title":"Force"},"description":"True=忽略已有结果重新生成"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ArticleDetailOut":{"properties":{"version_id":{"type":"integer","title":"Version Id"},"article_id":{"type":"integer","title":"Article Id"},"level_code":{"type":"string","title":"Level Code"},"lang":{"type":"string","title":"Lang"},"level":{"type":"integer","title":"Level"},"level_label":{"type":"string","title":"Level Label"},"title":{"type":"string","title":"Title"},"topic":{"type":"string","title":"Topic"},"source":{"type":"string","title":"Source"},"source_url":{"type":"string","title":"Source Url"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"published_date":{"type":"string","format":"date","title":"Published Date"},"published_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Published At"},"word_count":{"type":"integer","title":"Word Count","default":0},"reading_minutes":{"type":"number","title":"Reading Minutes","default":0.0},"lead":{"type":"string","title":"Lead","default":""},"preview":{"type":"string","title":"Preview","default":""},"body":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Body"},"vocab":{"items":{"$ref":"#/components/schemas/VocabWord"},"type":"array","title":"Vocab"},"audio":{"anyOf":[{"$ref":"#/components/schemas/AudioOut"},{"type":"null"}]},"paragraphs":{"items":{"type":"string"},"type":"array","title":"Paragraphs"},"title_original":{"type":"string","title":"Title Original","default":""},"summary_original":{"type":"string","title":"Summary Original","default":""},"body_original":{"type":"string","title":"Body Original","default":""},"fetched_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Fetched At"}},"type":"object","required":["version_id","article_id","level_code","lang","level","level_label","title","topic","source","source_url","published_date"],"title":"ArticleDetailOut"},"ArticleSummaryOut":{"properties":{"version_id":{"type":"integer","title":"Version Id"},"article_id":{"type":"integer","title":"Article Id"},"level_code":{"type":"string","title":"Level Code"},"lang":{"type":"string","title":"Lang"},"level":{"type":"integer","title":"Level"},"level_label":{"type":"string","title":"Level Label"},"title":{"type":"string","title":"Title"},"topic":{"type":"string","title":"Topic"},"source":{"type":"string","title":"Source"},"source_url":{"type":"string","title":"Source Url"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"published_date":{"type":"string","format":"date","title":"Published Date"},"published_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Published At"},"word_count":{"type":"integer","title":"Word Count","default":0},"reading_minutes":{"type":"number","title":"Reading Minutes","default":0.0},"lead":{"type":"string","title":"Lead","default":""},"preview":{"type":"string","title":"Preview","default":""},"body":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Body"},"vocab":{"items":{"$ref":"#/components/schemas/VocabWord"},"type":"array","title":"Vocab"},"audio":{"anyOf":[{"$ref":"#/components/schemas/AudioOut"},{"type":"null"}]}},"type":"object","required":["version_id","article_id","level_code","lang","level","level_label","title","topic","source","source_url","published_date"],"title":"ArticleSummaryOut","description":"列表项 —— 默认不含正文，避免 payload 过大。"},"AudioOut":{"properties":{"url":{"type":"string","title":"Url"},"duration":{"type":"number","title":"Duration"},"size_bytes":{"type":"integer","title":"Size Bytes","default":0},"engine":{"type":"string","title":"Engine","default":"edge-tts"},"voice":{"type":"string","title":"Voice","default":""},"has_timeline":{"type":"boolean","title":"Has Timeline","default":false},"word_count":{"type":"integer","title":"Word Count","default":0}},"type":"object","required":["url","duration"],"title":"AudioOut"},"FeedOut":{"properties":{"total":{"type":"integer","title":"Total"},"count":{"type":"integer","title":"Count"},"limit":{"type":"integer","title":"Limit"},"offset":{"type":"integer","title":"Offset"},"items":{"items":{"$ref":"#/components/schemas/ArticleSummaryOut"},"type":"array","title":"Items"}},"type":"object","required":["total","count","limit","offset","items"],"title":"FeedOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthOut":{"properties":{"status":{"type":"string","title":"Status"},"service":{"type":"string","title":"Service"},"version":{"type":"string","title":"Version"},"llm_provider":{"type":"string","title":"Llm Provider"},"llm_enabled":{"type":"boolean","title":"Llm Enabled"},"tts_enabled":{"type":"boolean","title":"Tts Enabled"},"tts_engine":{"type":"string","title":"Tts Engine"},"articles":{"type":"integer","title":"Articles"},"versions":{"type":"integer","title":"Versions"},"audios":{"type":"integer","title":"Audios"},"server_time":{"type":"string","title":"Server Time"}},"type":"object","required":["status","service","version","llm_provider","llm_enabled","tts_enabled","tts_engine","articles","versions","audios","server_time"],"title":"HealthOut"},"JobOut":{"properties":{"id":{"type":"integer","title":"Id"},"job":{"type":"string","title":"Job"},"status":{"type":"string","title":"Status"},"started_at":{"type":"string","format":"date-time","title":"Started At"},"finished_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Finished At"},"stats":{"additionalProperties":true,"type":"object","title":"Stats"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["id","job","status","started_at"],"title":"JobOut"},"LevelOut":{"properties":{"code":{"type":"string","title":"Code"},"lang":{"type":"string","title":"Lang"},"label":{"type":"string","title":"Label"},"level":{"type":"integer","title":"Level"},"target_words":{"type":"integer","title":"Target Words","default":0},"sentence_hint":{"type":"string","title":"Sentence Hint","default":""}},"type":"object","required":["code","lang","label","level"],"title":"LevelOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VocabWord":{"properties":{"word":{"type":"string","title":"Word"},"pos":{"type":"string","title":"Pos","default":""},"zh":{"type":"string","title":"Zh","default":""},"note":{"type":"string","title":"Note","default":""}},"type":"object","required":["word"],"title":"VocabWord"}}}}